Move coord vector
